Jira Search prototyping
Add credentials using previous projects as a guide
Add search that may eventually override this hardcoded jql. Took forever to build the url correctly as I had failed to remember I had https:// in the hostname or direnv wasn't exposing the right variables. Now we're returning results and we can likely start mapping to a response shape we wish to show. We likely need to take a step back to look at the design again.