Documentation merged with Readme
Moved documentation index up to the README as this should be our real documentation entry point.
I waffle on whether I prefer a compact readme vs having most of the documentation there but I think it helps to have a more robust readme and a separate file as things spill over.